+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
 # *********************************************************************************
-# URBANopt, Copyright (c) 2019-2020, Alliance for Sustainable Energy, LLC, and other
+# URBANopt (tm), Copyright (c) 2019-2020, Alliance for Sustainable Energy, LLC, and other
 # contributors. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without modification,
@@ -30,7 +30,7 @@
 
 require 'bundler/setup'
 require 'urbanopt/reopt'
-require 'urbanopt/scenario/default_reports'
+require 'urbanopt/reporting/default_reports'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure